fix: cleaning code
encapsuled the card method of sending and receiving money

+++ money-manager/src/main/java/com/demo/dto/InternTransferRequestDTO.java
@@ -1,4 +1,4 @@
 package com.demo.dto;
 
-public record InternTransferRequestDTO(Long id, Double amount, Long walletSenderId, Long walletReceiverId, Long cardSenderId, Long cardReceiverId) {
+public record InternTransferRequestDTO(Long id, float amount, Long walletSenderId, Long walletReceiverId, Long cardSenderId, Long cardReceiverId) {
 }
